+++ src/css/base.css
@@ -1,8 +1,11 @@
+html {
+    scroll-behavior: smooth;
+}
+
 body {
     color: var(--color);
     font-size: var(--font-size);
     font-family: var(--font-family);
-    scroll-behavior: smooth;
 }
 
 label {
